Reorganized build files; added stub publish script
