+++ scripts/release/package.json
@@ -14,6 +14,7 @@
     "cli-spinners": "^1.1.0",
     "command-line-args": "^4.0.7",
     "command-line-usage": "^4.0.1",
+    "figlet": "^1.2.0",
     "fs-extra": "^4.0.2",
     "log-update": "^2.1.0",
     "request-promise-json": "^1.0.4"
